feat(index.js): add option to clear to start
now if you call `clear({toStart: true})` it will clear to start (on my mac terminal at least), so that you can't scroll back to the cleared output.  When I'm running tests, linting, etc. in watch mode, I find it helpful to be able to nuke the old output so I can navigate through the latest errors as quickly as possible.